fix: head queries table
