+++ frontends/dashboard/src/analytics/components/charts/HeadQueries.tsx
@@ -1,9 +1,12 @@
 import { AnalyticsFilter, HeadQuery } from "shared/types";
-import { createMemo, Show } from "solid-js";
+import { createEffect, createMemo, Show, useContext } from "solid-js";
 import { SortableColumnDef, TanStackTable } from "shared/ui";
-import { useHeadQueries } from "../../hooks/data/useHeadQueries";
 import { createSolidTable, getCoreRowModel } from "@tanstack/solid-table";
 import { MagicSuspense } from "../../../components/MagicBox";
+import { DatasetContext } from "../../../contexts/DatasetContext";
+import { usePagination } from "../../hooks/usePagination";
+import { createQuery, useQueryClient } from "@tanstack/solid-query";
+import { getHeadQueries } from "../../api/analytics";
 
 interface HeadQueriesProps {
   params: { filter: AnalyticsFilter };
@@ -21,14 +24,52 @@ const columns: SortableColumnDef<HeadQuery>[] = [
 ];
 
 export const HeadQueries = (props: HeadQueriesProps) => {
-  const headQueriesData = createMemo(() => {
-    return useHeadQueries({
-      params: props.params,
+  const dataset = useContext(DatasetContext);
+  const pages = usePagination();
+  const queryClient = useQueryClient();
+
+  createEffect(() => {
+    // Preload the next page
+    const datasetId = dataset.datasetId();
+    const curPage = pages.page();
+    void queryClient.prefetchQuery({
+      queryKey: [
+        "head-queries",
+        { filters: props.params.filter, page: curPage + 1, dataset: datasetId },
+      ],
+      queryFn: async () => {
+        const results = await getHeadQueries(
+          props.params.filter,
+          datasetId,
+          curPage + 1,
+        );
+        if (results.length === 0) {
+          pages.setMaxPageDiscovered(curPage);
+        }
+        return results;
+      },
     });
   });
 
+  const headQueriesQuery = createQuery(() => ({
+    queryKey: [
+      "head-queries",
+      {
+        filters: props.params.filter,
+        page: pages.page(),
+        dataset: dataset.datasetId(),
+      },
+    ],
+    queryFn: () => {
+      return getHeadQueries(
+        props.params.filter,
+        dataset.datasetId(),
+        pages.page(),
+      );
+    },
+  }));
+
   const tableMemo = createMemo(() => {
-    const { headQueriesQuery, pages } = headQueriesData();
     const table = createSolidTable({
       get data() {
         return headQueriesQuery.data || [];
@@ -51,17 +92,9 @@ export const HeadQueries = (props: HeadQueriesProps) => {
     <MagicSuspense unstyled skeletonKey="headqueries">
       <Show
         fallback={<div class="py-8 text-center">No Data.</div>}
-        when={
-          headQueriesData()?.headQueriesQuery.data &&
-          headQueriesData().headQueriesQuery.data?.length
-        }
+        when={headQueriesQuery.data && headQueriesQuery.data?.length}
       >
-        <TanStackTable
-          small
-          pages={headQueriesData().pages}
-          perPage={10}
-          table={tableMemo()}
-        />
+        <TanStackTable small pages={pages} perPage={10} table={tableMemo()} />
       </Show>
     </MagicSuspense>
   );
